+++ lib/swarm/tracker/tracker.ex
@@ -692,7 +692,7 @@ defmodule Swarm.Tracker do
     debug "topology change (#{type} for #{remote_node})"
     current_node = state.self
     new_clock = Registry.reduce(state.clock, fn
-      entry(name: name, pid: pid, meta: meta) = obj, lclock when node(pid) == current_node ->
+      entry(name: name, pid: pid, meta: %{mfa: _mfa} = meta) = obj, lclock when node(pid) == current_node ->
         case Strategy.key_to_node(state.strategy, name) do
           :undefined ->
             # No node available to host process, it must be stopped
@@ -734,7 +734,7 @@ defmodule Swarm.Tracker do
                 lclock
             end
         end
-      entry(name: name, pid: pid, meta: meta) = obj, lclock when is_map(meta) ->
+      entry(name: name, pid: pid, meta: %{mfa: _mfa} = meta) = obj, lclock when is_map(meta) ->
         cond do
           Enum.member?(state.nodes, node(pid)) ->
             # the parent node is still up